# docs/openapi.yaml
openapi: 3.0.1
info:
  title: MTGLibrary API
  version: '1.0'
  description: API for MTGLibrary — endpoints for cards, collection management, decks, and users.
servers:
  - url: http://localhost:3000
paths:
...
  /cards/autocomplete:
    get:
      summary: Autocomplete card names
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Partial name to search for
      responses:
        '200':
          description: List of matching card names
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
  /cards/scryfall/{scryfallId}:
    get:
      summary: Lookup a card by Scryfall ID
      parameters:
        - in: path
          name: scryfallId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card found
        '404':
          description: Not found
  /cards/add:
    post:
      summary: Search for cards to add (by Name)
      description: Search global cards table by name to find candidates for your collection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                exact:
                  type: boolean
              required: [name]
      responses:
        '200':
          description: List of matches
  /cards/search:
    post:
      summary: Advanced Card Search
      description: Search with complex filters (color, type, stats, etc.)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                colors:
                  type: array
                  items:
                    type: string
                type:
                  type: string
                mv:
                  type: object
                  properties:
                    operator:
                      type: string
                    value:
                      type: number
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'

  # --- COLLECTION (User's Cards) ---
  /collection:
    get:
      summary: List User Collection
      parameters:
        - in: query
          name: deck_id
          schema:
            type: string
          description: Filter by deck ('null' or 'binder' for unassigned)
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: type_line
          schema:
            type: string
      responses:
        '200':
          description: List of user cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCard'
    post:
      summary: Add Single Card to Collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCardInput'
      responses:
        '201':
          description: Created
...
  /collection/batch:
    post:
      summary: Batch Import Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [merge, replace]
                cards:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserCardInput'
              required: [cards]
      responses:
        '200':
          description: Import result
  /collection/{id}:
    put:
      summary: Update User Card (Count, Deck, Finish)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                deck_id:
                  type: integer
                  nullable: true
                finish:
                  type: string
                is_wishlist:
                  type: boolean
      responses:
        '200':
          description: Updated
    delete:
      summary: Remove Card from Collection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted

  # --- DECKS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Card:
      type: object
      properties:
        name:
          type: string
        scryfall_id:
          type: string
        data:
          type: object
    UserCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        count:
          type: integer
        deck_id:
          type: integer
          nullable: true
        is_wishlist:
          type: boolean
        data:
          type: object
    UserCardInput:
      type: object
      properties:
        scryfall_id:
          type: string
        name:
          type: string
        count:
          type: integer
        finish:
          type: string
        deck_id:
          type: integer
        is_wishlist:
          type: boolean
        data:
          type: object